First whack at implementing timeliner, but it's a mess, screen no longer works, and tests are now broken

+++ build.sh
@@ -29,6 +29,13 @@ fi
     --target all \
     -j 10 \
 
+STATUS=$?
+if [ $STATUS -ne 0 ]; then
+    echo "CMake returned error exit code: ${STATUS}"
+    echo "Exiting"
+    exit $STATUS
+fi
+
 mkdir -p ${DIST_DIR}
 rm -rf ${DIST_DIR}/*
 cp ${BUILD_DIR}/src/main/*.uf2 ${DIST_DIR}
